Reorganize backtests into venue tick folders
