+++ AGENTS.md
@@ -20,7 +20,7 @@ Useful smoke checks:
 ```bash
 uv run python backtests/kalshi_trade_tick/kalshi_breakout.py
 uv run python backtests/polymarket_trade_tick/polymarket_vwap_reversion.py
-uv run python backtests/polymarket_pmxt_relay_ema_crossover.py
+uv run python backtests/polymarket_quote_tick/polymarket_pmxt_relay_ema_crossover.py
 ```
 
 ## What Matters Most
