+++ backtests/_shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared backtest-layer helpers."""
