+++ backtests/kalshi_quote_tick/__init__.py
@@ -1,0 +1,4 @@
+"""Kalshi quote-tick backtests.
+
+This package is intentionally empty for now.
+"""
