+++ backtests/kalshi_trade_tick/_defaults.py
@@ -1,1 +1,12 @@
-from backtests._defaults import *  # noqa: F403
+# Derived from NautilusTrader prediction-market example code.
+# Distributed under the GNU Lesser General Public License Version 3.0 or later.
+# Modified in this repository on 2026-03-27.
+# See the repository NOTICE file for provenance and licensing scope.
+
+"""
+Defaults for Kalshi trade-tick backtests.
+"""
+
+DEFAULT_KALSHI_MARKET_TICKER = "KXNEXTIRANLEADER-45JAN01-MKHA"
+DEFAULT_LOOKBACK_DAYS = 30
+DEFAULT_INITIAL_CASH = 100.0
