+++ backtests/polymarket_quote_tick/__init__.py
@@ -1,0 +1,1 @@
+"""Polymarket quote-tick backtests backed by PMXT historical L2 data."""
