+++ backtests/polymarket_trade_tick/__init__.py
@@ -1,6 +1,5 @@
 """Archived Polymarket trade-tick backtests.
 
-These legacy runners are kept for reference and direct use, but they are not
-part of the default public backtest menu anymore. The public Polymarket path is
-the PMXT relay-backed `backtests/polymarket_pmxt_relay_*.py` set.
+These legacy runners are kept for reference and direct use alongside the PMXT
+relay-backed `backtests/polymarket_quote_tick/` set.
 """
