+++ backtests/polymarket_trade_tick/_defaults.py
@@ -1,1 +1,14 @@
-from backtests._defaults import *  # noqa: F403
+# Derived from NautilusTrader prediction-market example code.
+# Distributed under the GNU Lesser General Public License Version 3.0 or later.
+# Modified in this repository on 2026-03-27.
+# See the repository NOTICE file for provenance and licensing scope.
+
+"""
+Defaults for Polymarket trade-tick backtests.
+"""
+
+DEFAULT_POLYMARKET_MARKET_SLUG = (
+    "will-openai-launch-a-new-consumer-hardware-product-by-march-31-2026"
+)
+DEFAULT_LOOKBACK_DAYS = 30
+DEFAULT_INITIAL_CASH = 100.0
