+++ backtests/polymarket_trade_tick/polymarket_sports_final_period_momentum.py
@@ -20,11 +20,9 @@ from __future__ import annotations
 
 import asyncio
 import os
-import sys
 from datetime import UTC
 from datetime import datetime
 from decimal import Decimal
-from pathlib import Path
 
 try:
     from ._script_helpers import ensure_repo_root
@@ -58,18 +56,17 @@ from strategies import (
 
 
 try:
-    from _defaults import DEFAULT_LOOKBACK_DAYS
-    from _polymarket_single_market_runner import run_single_market_trade_backtest
-    from _script_helpers import parse_bool_env
-    from _script_helpers import parse_csv_env
-except ModuleNotFoundError:
-    _THIS_DIR = Path(__file__).resolve().parent
-    if str(_THIS_DIR) not in sys.path:
-        sys.path.insert(0, str(_THIS_DIR))
-    from _defaults import DEFAULT_LOOKBACK_DAYS
-    from _polymarket_single_market_runner import run_single_market_trade_backtest
-    from _script_helpers import parse_bool_env
-    from _script_helpers import parse_csv_env
+    from ._defaults import DEFAULT_LOOKBACK_DAYS
+    from ._polymarket_single_market_runner import run_single_market_trade_backtest
+    from ._script_helpers import parse_bool_env
+    from ._script_helpers import parse_csv_env
+except ImportError:
+    from backtests.polymarket_trade_tick._defaults import DEFAULT_LOOKBACK_DAYS
+    from backtests.polymarket_trade_tick._polymarket_single_market_runner import (
+        run_single_market_trade_backtest,
+    )
+    from backtests.polymarket_trade_tick._script_helpers import parse_bool_env
+    from backtests.polymarket_trade_tick._script_helpers import parse_csv_env
 
 
 NAME = "polymarket_sports_final_period_momentum"
