+++ backtests/polymarket_trade_tick/polymarket_vwap_reversion.py
@@ -13,9 +13,7 @@ from __future__ import annotations
 
 import asyncio
 import os
-import sys
 from decimal import Decimal
-from pathlib import Path
 
 try:
     from ._script_helpers import ensure_repo_root
@@ -29,18 +27,17 @@ from strategies import TradeTickVWAPReversionStrategy
 
 
 try:
-    from _defaults import DEFAULT_INITIAL_CASH
-    from _defaults import DEFAULT_LOOKBACK_DAYS
-    from _defaults import DEFAULT_POLYMARKET_MARKET_SLUG
-    from _polymarket_single_market_runner import run_single_market_trade_backtest
-except ModuleNotFoundError:
-    _THIS_DIR = Path(__file__).resolve().parent
-    if str(_THIS_DIR) not in sys.path:
-        sys.path.insert(0, str(_THIS_DIR))
-    from _defaults import DEFAULT_INITIAL_CASH
-    from _defaults import DEFAULT_LOOKBACK_DAYS
-    from _defaults import DEFAULT_POLYMARKET_MARKET_SLUG
-    from _polymarket_single_market_runner import run_single_market_trade_backtest
+    from ._defaults import DEFAULT_INITIAL_CASH
+    from ._defaults import DEFAULT_LOOKBACK_DAYS
+    from ._defaults import DEFAULT_POLYMARKET_MARKET_SLUG
+    from ._polymarket_single_market_runner import run_single_market_trade_backtest
+except ImportError:
+    from backtests.polymarket_trade_tick._defaults import DEFAULT_INITIAL_CASH
+    from backtests.polymarket_trade_tick._defaults import DEFAULT_LOOKBACK_DAYS
+    from backtests.polymarket_trade_tick._defaults import DEFAULT_POLYMARKET_MARKET_SLUG
+    from backtests.polymarket_trade_tick._polymarket_single_market_runner import (
+        run_single_market_trade_backtest,
+    )
 
 
 NAME = "polymarket_vwap_reversion"
